Add ghost pattern and ghost factory

+++ code/include/utils/Vec2.hpp
@@ -2,27 +2,27 @@
 
 class Vec2 {
 public:
-    float x, y;
+    int x, y;
 
     Vec2() : x(0.f), y(0.f) {}
-    Vec2(float x_, float y_) : x(x_), y(y_) {}
+    Vec2(int x_, int y_) : x(x_), y(y_) {}
 
     Vec2 operator+(const Vec2& other) const;
     Vec2 operator-(const Vec2& other) const;
-    Vec2 operator*(float scalar) const;
-    Vec2 operator/(float scalar) const;
+    Vec2 operator*(int scalar) const;
+    Vec2 operator/(int scalar) const;
 
     Vec2& operator+=(const Vec2& other);
     Vec2& operator-=(const Vec2& other);
     Vec2& operator*=(const Vec2& other);
-    Vec2& operator*=(float scalar);
-    Vec2& operator/=(float scalar);
+    Vec2& operator*=(int scalar);
+    Vec2& operator/=(int scalar);
 
     bool operator!=(const Vec2& other) const;
     bool operator==(const Vec2& other) const;
 
-    float Length() const;
-    float LengthSquared() const;
+    int Length() const;
+    int LengthSquared() const;
     Vec2 Normalized() const;
-    float Dot(const Vec2& other) const;
+    int Dot(const Vec2& other) const;
 };
